リソースがない時の処理とバリデーションエラーの処理方法を変更

+++ src/main/java/com/example/demo/service/AnimeService.java
@@ -1,6 +1,7 @@
 package com.example.demo.service;
 
 import com.example.demo.entity.Anime;
+import com.example.demo.exception.ResourceNotFoundException;
 import com.example.demo.mapper.AnimeMapper;
 import org.springframework.stereotype.Service;
 
@@ -21,9 +22,12 @@ public class AnimeService {
         return animeList;
     }
 
-    public Optional<Anime> getAnime(Integer id) {
+    public Anime getAnime(Integer id) {
         Optional<Anime> animeId = animeMapper.findById(id);
-        return animeId;
+        if (animeId.isPresent()) {
+            return animeId.get();
+        }
+        throw new ResourceNotFoundException("resource not found");
     }
 
     public void registerAnime(String name, String genre) {
@@ -32,11 +36,19 @@ public class AnimeService {
     }
 
     public void updateAnime(Integer id, String name, String genre) {
-        Anime anime = new Anime(id, name, genre);
-        animeMapper.updateAnime(anime);
+        Optional<Anime> animeId = animeMapper.findById(id);
+        if (animeId.isPresent()) {
+            Anime anime = new Anime(id, name, genre);
+            animeMapper.updateAnime(anime);
+        }
+        throw new ResourceNotFoundException("resource not found");
     }
 
     public void deleteAnime(Integer id) {
-        animeMapper.deleteAnime(id);
+        Optional<Anime> animeId = animeMapper.findById(id);
+        if (animeId.isPresent()) {
+            animeMapper.deleteAnime(id);
+        }
+        throw new ResourceNotFoundException("resource not found");
     }
 }
